indicators: Candle Size Multiplier must be an int, not float

+++ indicators.py
@@ -101,11 +101,11 @@ class Helpers:
 # Relative Strength Index
 class RSI:
   CandleDepends = (int(config.gc['Indicators']['RSI'][
-                   'Period']) * float(config.gc['Indicators']['RSI']['Candle Size Multiplier'])) + 1
+                   'Period']) * int(config.gc['Indicators']['RSI']['Candle Size Multiplier'])) + 1
 
   def indicator():
     Period = int(config.gc['Indicators']['RSI'][
-                 'Period']) * float(config.gc['Indicators']['RSI']['Candle Size Multiplier'])
+                 'Period']) * int(config.gc['Indicators']['RSI']['Candle Size Multiplier'])
     # We need a minimum of 2 candles to start RS calculations
     if len(ldb.price_list) >= 2:
       if ldb.price_list[-1] > ldb.price_list[-2]:
@@ -153,13 +153,13 @@ class RSI:
 # Simple Movement Average
 class SMA:
   CandleDepends = int(config.gc['Indicators']['SMA'][
-                      'Long Period']) * float(config.gc['Indicators']['SMA']['Candle Size Multiplier'])
+                      'Long Period']) * int(config.gc['Indicators']['SMA']['Candle Size Multiplier'])
 
   def indicator():
     LongPeriod = int(config.gc['Indicators']['SMA'][
-                     'Long Period']) * float(config.gc['Indicators']['SMA']['Candle Size Multiplier'])
+                     'Long Period']) * int(config.gc['Indicators']['SMA']['Candle Size Multiplier'])
     ShortPeriod = int(config.gc['Indicators']['SMA'][
-                      'Short Period']) * float(config.gc['Indicators']['SMA']['Candle Size Multiplier'])
+                      'Short Period']) * int(config.gc['Indicators']['SMA']['Candle Size Multiplier'])
     # We can start SMA calculations once we have max period candles
     if len(ldb.price_list) >= max(LongPeriod, ShortPeriod):
       storage.writelist(
@@ -184,13 +184,13 @@ class SMA:
 # Exponential Movement Average
 class EMA:
   CandleDepends = int(config.gc['Indicators']['EMA'][
-                      'Long Period']) * float(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
+                      'Long Period']) * int(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
 
   def indicator():
     LongPeriod = int(config.gc['Indicators']['EMA'][
-                     'Long Period']) * float(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
+                     'Long Period']) * int(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
     ShortPeriod = int(config.gc['Indicators']['EMA'][
-                      'Short Period']) * float(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
+                      'Short Period']) * int(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
     # We can start EMAs once we have max period candles
     if len(ldb.price_list) >= max(LongPeriod, ShortPeriod):
       storage.writelist('EMA_Short_list', Helpers.EMA(
@@ -214,14 +214,14 @@ class EMA:
 class DEMA:
   CandleDepends = int(config.gc['Indicators']['EMA'][
                       'Long Period']) + (int(config.gc[
-                          'Indicators']['EMA']['Short Period']) * 2) * float(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
+                          'Indicators']['EMA']['Short Period']) * 2) * int(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
   IndicatorDepends = ['EMA']
 
   def indicator():
     LongPeriod = int(config.gc['Indicators']['EMA'][
-                     'Long Period']) * float(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
+                     'Long Period']) * int(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
     ShortPeriod = int(config.gc['Indicators']['EMA'][
-                      'Short Period']) * float(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
+                      'Short Period']) * int(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
     # We can start DEMAs once we have max period candles
     if len(storage.getlist('EMA_Long_list')) >= max(LongPeriod, ShortPeriod):
       storage.writelist('DEMA_Short_list', Helpers.DEMA(storage.getlist(
@@ -244,11 +244,11 @@ class DEMA:
 # Exponential Movement Average (using wbic16's logic)
 class EMAwbic:
   CandleDepends = int(config.gc['Indicators']['EMAwbic'][
-                      'Period']) * float(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
+                      'Period']) * int(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
 
   def indicator():
     Period = int(config.gc['Indicators']['EMA'][
-                 'Period']) * float(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
+                 'Period']) * int(config.gc['Indicators']['EMA']['Candle Size Multiplier'])
     if len(ldb.price_list) >= Period:
       storage.writelist('EMAwbic_EMA_list', Helpers.EMA(
           ldb.price_list, storage.getlist('EMAwbic_EMA_list'), Period))
@@ -265,13 +265,13 @@ class EMAwbic:
 # Fractal Adaptive Moving Average
 class FRAMA:
   CandleDepends = int(config.gc['Indicators']['FRAMA'][
-                      'Long Period']) * float(config.gc['Indicators']['FRAMA']['Candle Size Multiplier'])
+                      'Long Period']) * int(config.gc['Indicators']['FRAMA']['Candle Size Multiplier'])
 
   def indicator():
     LongPeriod = int(config.gc['Indicators']['FRAMA'][
-                     'Long Period']) * float(config.gc['Indicators']['FRAMA']['Candle Size Multiplier'])
+                     'Long Period']) * int(config.gc['Indicators']['FRAMA']['Candle Size Multiplier'])
     ShortPeriod = int(config.gc['Indicators']['FRAMA'][
-                      'Short Period']) * float(config.gc['Indicators']['FRAMA']['Candle Size Multiplier'])
+                      'Short Period']) * int(config.gc['Indicators']['FRAMA']['Candle Size Multiplier'])
     # We can start FRAMAs once we have max period candles
     if len(ldb.price_list) >= (max(LongPeriod, ShortPeriod)):
       try:
@@ -299,13 +299,13 @@ class FRAMA:
 # Movement Average Convergence Divergence
 class MACD:
   CandleDepends = int(config.gc['Indicators']['MACD']['Long Period']) + (int(
-      config.gc['Indicators']['MACD']['Short Period']) / 2) * float(config.gc['Indicators']['MACD']['Candle Size Multiplier'])
+      config.gc['Indicators']['MACD']['Short Period']) / 2) * int(config.gc['Indicators']['MACD']['Candle Size Multiplier'])
 
   def indicator():
     LongPeriod = int(config.gc['Indicators']['MACD'][
-                     'Long Period']) * float(config.gc['Indicators']['MACD']['Candle Size Multiplier'])
+                     'Long Period']) * int(config.gc['Indicators']['MACD']['Candle Size Multiplier'])
     ShortPeriod = int(config.gc['Indicators']['MACD'][
-                      'Short Period']) * float(config.gc['Indicators']['MACD']['Candle Size Multiplier'])
+                      'Short Period']) * int(config.gc['Indicators']['MACD']['Candle Size Multiplier'])
     SignalPeriod = int(config.gc['Indicators']['MACD']['Signal Period'])
     # We can start MACD EMAs once we have max period candles
     if len(ldb.price_list) >= max(LongPeriod, ShortPeriod):
@@ -338,13 +338,13 @@ class MACD:
 class DMACD:
   IndicatorDepends = ['MACD']
   CandleDepends = ((int(config.gc['Indicators']['MACD']['Long Period']) + (int(
-      config.gc['Indicators']['MACD']['Short Period']) / 2)) * 2) * float(config.gc['Indicators']['MACD']['Candle Size Multiplier'])
+      config.gc['Indicators']['MACD']['Short Period']) / 2)) * 2) * int(config.gc['Indicators']['MACD']['Candle Size Multiplier'])
 
   def indicator():
     LongPeriod = int(config.gc['Indicators']['MACD'][
-                     'Long Period']) * float(config.gc['Indicators']['MACD']['Candle Size Multiplier'])
+                     'Long Period']) * int(config.gc['Indicators']['MACD']['Candle Size Multiplier'])
     ShortPeriod = int(config.gc['Indicators']['MACD'][
-                      'Short Period']) * float(config.gc['Indicators']['MACD']['Candle Size Multiplier'])
+                      'Short Period']) * int(config.gc['Indicators']['MACD']['Candle Size Multiplier'])
     SignalPeriod = int(config.gc['Indicators']['MACD']['Signal Period'])
     # We can start DEMAs once we have max period candles
     if len(storage.getlist('MACD_Long_list')) >= max(LongPeriod, ShortPeriod):
@@ -376,11 +376,11 @@ class DMACD:
 # Fast Stochastic %K
 class FastStochK:
   CandleDepends = int(config.gc['Indicators']['Fast Stochastic %K'][
-                      'Period']) * float(config.gc['Indicators']['Fast Stochastic %K']['Candle Size Multiplier'])
+                      'Period']) * int(config.gc['Indicators']['Fast Stochastic %K']['Candle Size Multiplier'])
 
   def indicator():
     Period = int(config.gc['Indicators']['Fast Stochastic %K'][
-                 'Period']) * float(config.gc['Indicators']['Fast Stochastic %K']['Candle Size Multiplier'])
+                 'Period']) * int(config.gc['Indicators']['Fast Stochastic %K']['Candle Size Multiplier'])
     # We can start FastStochK calculations once we have FastStochKPeriod
     # candles, otherwise we append None until met
     if len(ldb.price_list) >= Period:
@@ -517,15 +517,15 @@ class FullStochRSID:
 class KDJ:
   CandleDepends = int(config.gc['Indicators']['KDJ']['Fast K Period']) + int(
       config.gc['Indicators']['KDJ']['Full K Period']) + (int(config.gc[
-          'Indicators']['KDJ']['Full D Period']) - 2) * float(config.gc['Indicators']['KDJ']['Candle Size Multiplier'])
+          'Indicators']['KDJ']['Full D Period']) - 2) * int(config.gc['Indicators']['KDJ']['Candle Size Multiplier'])
 
   def indicator():
     FastKPeriod = int(config.gc['Indicators']['KDJ'][
-                      'Fast K Period']) * float(config.gc['Indicators']['KDJ']['Candle Size Multiplier'])
+                      'Fast K Period']) * int(config.gc['Indicators']['KDJ']['Candle Size Multiplier'])
     FullKPeriod = int(config.gc['Indicators']['KDJ'][
-                      'Full K Period']) * float(config.gc['Indicators']['KDJ']['Candle Size Multiplier'])
+                      'Full K Period']) * int(config.gc['Indicators']['KDJ']['Candle Size Multiplier'])
     FullDPeriod = int(config.gc['Indicators']['KDJ'][
-                      'Full D Period']) * float(config.gc['Indicators']['KDJ']['Candle Size Multiplier'])
+                      'Full D Period']) * int(config.gc['Indicators']['KDJ']['Candle Size Multiplier'])
     if len(ldb.price_list) >= FastKPeriod:
       try:
         storage.writelist(
@@ -555,11 +555,11 @@ class KDJ:
 # Aroon Oscillator
 class Aroon:
   CandleDepends = int(config.gc['Indicators']['Aroon'][
-                      'Period']) * float(config.gc['Indicators']['Aroon']['Candle Size Multiplier'])
+                      'Period']) * int(config.gc['Indicators']['Aroon']['Candle Size Multiplier'])
 
   def indicator():
     Period = int(config.gc['Indicators']['Aroon'][
-                 'Period']) * float(config.gc['Indicators']['Aroon']['Candle Size Multiplier'])
+                 'Period']) * int(config.gc['Indicators']['Aroon']['Candle Size Multiplier'])
     # We must have AroonPeriod ldb.price_list candles
     if len(ldb.price_list) >= Period:
       storage.writelist('Aroon_Up_list', 100 * (Period -
@@ -591,13 +591,13 @@ class Ichimoku:
 
   def indicator():
     TSP = int(config.gc['Indicators']['Ichimoku']['Tenkan-Sen Period']) * \
-        float(config.gc['Indicators']['Ichimoku']['Candle Size Multiplier'])
+        int(config.gc['Indicators']['Ichimoku']['Candle Size Multiplier'])
     SSP = int(config.gc['Indicators']['Ichimoku']['Senkou Span Period']) * \
-        float(config.gc['Indicators']['Ichimoku']['Candle Size Multiplier'])
+        int(config.gc['Indicators']['Ichimoku']['Candle Size Multiplier'])
     KSP = int(config.gc['Indicators']['Ichimoku']['Kijun-Sen Period']) * \
-        float(config.gc['Indicators']['Ichimoku']['Candle Size Multiplier'])
+        int(config.gc['Indicators']['Ichimoku']['Candle Size Multiplier'])
     CSP = int(config.gc['Indicators']['Ichimoku']['Chikou Span Period']) * \
-        float(config.gc['Indicators']['Ichimoku']['Candle Size Multiplier'])
+        int(config.gc['Indicators']['Ichimoku']['Candle Size Multiplier'])
     IS = config.gc['Indicators']['Ichimoku']['Indicator Strategy']
     # We must have SenkouSpanPeriod price candles before starting
     # calculations, otherwise we append None
@@ -747,7 +747,7 @@ class StdDev:
 
   def indicator():
     Period = int(config.gc['Indicators']['Standard Deviation'][
-                 'Period']) * float(config.gc['Indicators']['Standard Deviation']['Candle Size Multiplier'])
+                 'Period']) * int(config.gc['Indicators']['Standard Deviation']['Candle Size Multiplier'])
     # We can start StdDev calculations once we have StdDevSample
     # candles, otherwise we append None until met
     if len(ldb.price_list) >= Period:
@@ -764,7 +764,7 @@ class StdDev:
 # Bollinger Bands
 class BollBands:
   CandleDepends = int(config.gc['Indicators']['Bollinger Bands'][
-                      'Period']) * float(config.gc['Indicators']['Bollinger Bands']['Candle Size Multiplier'])
+                      'Period']) * int(config.gc['Indicators']['Bollinger Bands']['Candle Size Multiplier'])
 
   def indicator():
     Period = int(config.gc['Indicators']['Bollinger Bands']['Period'])
@@ -781,7 +781,7 @@ class BollBands:
 # Bollinger Bandwidth
 class BollBandwidth:
   CandleDepends = int(config.gc['Indicators']['Bollinger Bands'][
-                      'Period']) * float(config.gc['Indicators']['Bollinger Bands']['Candle Size Multiplier'])
+                      'Period']) * int(config.gc['Indicators']['Bollinger Bands']['Candle Size Multiplier'])
   IndicatorDepends = ['BollBands']
 
   def indicator():
@@ -800,11 +800,11 @@ class BollBandwidth:
 # Average True Range
 class ATR:
   CandleDepends = ((int(config.gc['Indicators']['Average True Range'][
-                   'Period']) * 3) - 1) * float(config.gc['Indicators']['Average True Range']['Candle Size Multiplier'])
+                   'Period']) * 3) - 1) * int(config.gc['Indicators']['Average True Range']['Candle Size Multiplier'])
 
   def indicator():
     Period = int(config.gc['Indicators']['Average True Range'][
-                 'Period']) * float(config.gc['Indicators']['Average True Range']['Candle Size Multiplier'])
+                 'Period']) * int(config.gc['Indicators']['Average True Range']['Candle Size Multiplier'])
     # We can start ATR calculations once we have two periods
     if len(ldb.price_list) >= (Period * 2):
       storage.writelist(
@@ -824,11 +824,11 @@ class ATR:
 class ChandExit:
   CandleDepends = ((int(config.gc['Indicators']['Chandelier Exit']['Period']) *
                     int(config.gc['Indicators']['Chandelier Exit']['Multiplier'])) - 1) * \
-                    float(config.gc['Indicators']['Ichimoku']['Candle Size Multiplier'])
+                    int(config.gc['Indicators']['Ichimoku']['Candle Size Multiplier'])
 
   def indicator():
     Period = int(config.gc['Indicators']['Chandelier Exit'][
-                 'Period']) * float(config.gc['Indicators']['Chandelier Exit']['Candle Size Multiplier'])
+                 'Period']) * int(config.gc['Indicators']['Chandelier Exit']['Candle Size Multiplier'])
     Multiplier = int(config.gc['Indicators']['Chandelier Exit']['Multiplier'])
     # We can start calculations once we have two periods
     if len(ldb.price_list) >= (Period * 2):
@@ -866,12 +866,12 @@ class ChandExit:
 # Directional Movement
 class DMI:
   CandleDepends = (int(config.gc['Indicators']['Average True Range'][
-                   'Period']) * 5) * float(config.gc['Indicators']['Average True Range']['Candle Size Multiplier'])
+                   'Period']) * 5) * int(config.gc['Indicators']['Average True Range']['Candle Size Multiplier'])
   DMITrend = 'No Trend'
 
   def indicator():
     Period = int(config.gc['Indicators']['Average True Range'][
-                 'Period']) * float(config.gc['Indicators']['Average True Range']['Candle Size Multiplier'])
+                 'Period']) * int(config.gc['Indicators']['Average True Range']['Candle Size Multiplier'])
     # We can start DMI calculations once we have two ATR periods
     if len(ldb.price_list) >= (Period * 2):
       UpMove = max(ldb.price_list[-Period:]) - max(
@@ -947,11 +947,11 @@ class DMI:
 # (Simple) Rate of Change (Momentum)
 class SROC:
   CandleDepends = (int(config.gc['Indicators']['Simple Rate of Change'][
-                   'Period']) + 1) * float(config.gc['Indicators']['Simple Rate of Change']['Candle Size Multiplier'])
+                   'Period']) + 1) * int(config.gc['Indicators']['Simple Rate of Change']['Candle Size Multiplier'])
 
   def indicator():
     Period = int(config.gc['Indicators']['Simple Rate of Change'][
-                 'Period']) * float(config.gc['Indicators']['Simple Rate of Change']['Candle Size Multiplier'])
+                 'Period']) * int(config.gc['Indicators']['Simple Rate of Change']['Candle Size Multiplier'])
     # We can start ROC calculations once we have SROC Periods of Price
     if len(ldb.price_list) >= Period:
       storage.writelist(
